Share seller product image path ownership logic
